The Synesthesia project is a little experiment of 3D and virtual reality code that question the link between objects and focus on senses in a VR environment.
The Synesthesia project is a little experiment of 3D and virtual reality code that question the link between objects and focus on senses in a VR environment.
The main idea was to create a visual zone where the player could paint and draw with moving "palettes" of colors. To add the synesthetic dimension, every color is associated to a sound. In that sense, the player is creating simultaneously a visual and auditive piece of pixel art. The "white cube" where the player can experiment is an addition of six matrices that can be used as painting canvas, from the inside as much as the outside of the cube. Consequently, every face of the cube can be used to express creativity and to try to mix visual and auditive senses to create something unique. The white color in the middle of the palette is a kind of reset to allows the player to wipe off the color and its sound, restart a new work, again and again.